requirements.txt
boto==2.38.0
chardet==2.3.0
cryptography==1.2.3
dj-dajaxice==1.0
dj-static==0.0.6
Django==1.9.5
django-annoying==0.10.3
django-bootstrap-form==3.2.1
django-dajax==0.9.2
django-dajaxice==0.7
django-floppyforms==1.7.0
django-jet==1.0.2
django-mathfilters==0.4.0
django-suit==0.2.21
docopt==0.6.2
duplicity==0.7.6
enum34==1.1.2
feedparser==5.2.1
gyp==0.1
idna==2.0
ipaddress==1.0.16
lockfile==0.12.2
ndg-httpsclient==0.4.0
pipreqs==0.4.4
pyasn1==0.1.9
pygobject==3.20.0
pyOpenSSL==0.15.1
python-cloudfiles==1.7.10
pyxdg==0.25
requests==2.9.1
six==1.10.0
static3==0.7.0
urllib3==1.13.1
yarg==0.1.9
